Proof of concept for Paper installer
Paper will be distributed as a zipapp which contains bootstrap
(vendored) wheels and a bootstrap script; the script creates a private
venv for the installed Paper code along with cache directories for both
packed and unpacked versions of wheels, installs its own wheels (by
hard-linking the unpacked files into its environment) and (TODO) sets
up the main Paper executable wrapper.

Thereafter, Paper should be able to run normally, taking advantage of
its bootstrapped dependencies (on first run, it may use these to grab
more dependencies: e.g. pre-resolved dependencies for a resolver, and
tools for providing an enhanced UI) which are also immediately available
in cache (i.e. can be installed in other environments without a
download, when requested).

The hard-link installation strategy (others may be implemented later)
allows environments to share files for distribution packages that they
have in common, while still appearing to be isolated. (This breaks down
for packages that expect to modify included data files.)

For testing purposes, the zipapp includes `package-installation-test` as
a bootstrap wheel (obviously, this is not a real Paper dependency).

The included `demo.sh` goes through a cycle of building the Paper wheel,
bundling it (via `make-zipapp`) in a zipapp with the bootstrap script
(`install.py`, renamed in the zipapp) and other wheels, and finally
running the zipapp.