Proof of concept for installation
The new code uses Pradyun Gedam's `installer` package to install into a
specified venv (i.e. in general, not the one Paper is running in) using
the standard Pip strategy (unpack files from the wheel, copy and make
script wrappers). It parses the venv's config data to determine
important paths.

The included `installer` wheel is stripped of Windows-specific wrapper
exectuables. Windows support will be figured out later, but the plan is
to have Windows installations of Paper download only the stub needed for
the current system.

(The final bootstrap installer will probably dynamically grab wheels
using raw `urllib`, and then delegate to the "core" installer wheel to
finish installation.)

For now, in testing, the `installer` wheel itself as well as the `src`
directory need to be on `sys.path` (e.g. by setting `PYTHONPATH`). The
bootstrap script should ensure everything is in the right place for an
installed version.

Upcoming tasks:

* Make Paper fetch a wheel from its own wheelhouse for installation
* Support other installation strategies (hard links, symlinks and .pth
  files referring to the unpacked cache
* Integrate other functional steps: downloading from PyPI (perhaps
  bootstrapped so that `requests` etc. isn't needed in the zipapp),
  building via `build` and resolving dependencies via `resolvelib`