Refactoring
Split the work into separate functions for clarity.
`importlib` is cheap to import, but `pathlib`, `subprocess` and
`configparser` are not (several ms each). These are therefore deferred
so that they don't automatically happen on runs where the script has to
restart itself with a different Python interpreter anyway. However, the
`pathlib` import is part of that detection logic, so it still happens.

The next step is to try to replace the `pathlib` and `subprocess` work
with their raw `os` equivalents.