Add all cache wheels to import path during install
Since the unpack logic has to iterate over the zipfile anyway, it simply
detects each wheel it unpacks and adds them to `sys.path`. This
simplifies the overall logic, since the firstrun code no longer has to
figure out where the `rfbs` wheel (or, for now, the `installer` wheel) is.

Theoretically, since the firstrun code will have to iterate to find all
the wheels to set up, it could add them to path as it goes. However, its
*own* wheel (i.e., the `paper` wheel) is already there, and the initial
code has to put it there. Overall, this way seems simpler.