Vendor `installer` (for now)
Rather than including a partial wheel for `installer` version 0.7.0 with
the windows stub exectuables removed (this approach was really broken,
since e.g. the RECORD file still listed those binary files), a patched
version of `installer` version 0.4.0 (the last claiming support for
Python 3.5) is included into the `rfbs` wheel.

In the long run, this code will be replaced anyway - to simplify and
streamline the installation process, and to make it handle installation
from a pre-unpacked cache.

The code is made into a sub-package so that Flit can still see a single
top-level package for the distribution. (`bbbb` won't have such a
limitation.) To make this work, a bunch of imports had to be fixed.

The Windows stub executables are again removed, and the code that would
have handled them has been patched to avoid the dependency on
`importlib_resources` for older Python versions that don't include
`importlib.resources` in the standard library. (When Windows support is
added, a workaround for missing `importlib.resources.files()` will be
added directly. This seems difficult, but entirely possible.)