Implement generic directory-tree visitor
This code generalizes the process of searching through a source folder
(for example, the package folders or `.dist-info` folder from an
unpacked wheel) and copying, hard-linking, symbolic-linking etc. the
contents to a parallel destination folder. This is crucial for setting
up packages from the cache.

This code is supposed to be an implementation detail for now, but may be
exposed (or moved to a separate package) in the future. Thus, a lengthy
comment has been included that could be converted to a docstring.