Fix bugs
The top-level folder also needs to be passed to `process_folder`; the
code is changed such that the current folder is processed inside the
`os.walk` loop rather than iterating over subdirectories. This fixes the
bug and also avoids a nested iteration (via list comprehension);
processing of contained files and folders is now all-or-nothing.

The documentation comment is updated to match, and a note is added about
the use of `onerror` vis-a-vis processing functions.