Implement "copy distribution into cache" logic
Wheels and sdists can now be copied into the appropriate subdirectory of
a PAPER cache, computed according to the distribution name and version
as represented by the filename. Some verification is done that the
filename matches modern standards, but the contents of the distribution
archive are not checked.

If a matching file is already present in the cache, copying will be
skipped unless `force=True` is provided.

Additionally, the code to find that folder in the cache will now use the
default cache location if `None` is provided - this logic has been
promoted from `setup_from_cache` because it's also needed here and will
likely benefit future code as well.