Flatten bundled wheel structure in zipapp
Instead of copying a pre-loaded cache into the zipapp, it now just
contains a folder with the wheels that will be loaded into the cache.
This change is important for several reasons:

* In case the cache structure changes in the future, this reduces the
  maintenance burden for the zipapp maker.

* The resulting zipapp can be much smaller. (This trades off against
  time for the initial installation.)

* Going forward, the actual cache folder location should be calculated
  by `rfbs`, so that it can do platform-specific things. That location
  may also have something there already, such as a previous PAPER cache
  in an older, incompatible version.

The new installation procedure will be to unpack the wheels into a
temporary directory, put all of them on the path, and then use the
`rfbs` API to ensure the initial cache location and store the initial
wheels there, using the new `copy_distribution_into_cache`. From there,
installation from cache into the venv can proceed normally, again using
the `rfbs` API.

Finally, a vendored `packaging` wheel is added, since it has been
discovered that `resolvelib` won't figure out what wheels or sdists to
download, only what versions to try. There are many other things in
`packaging` that are also likely to prove useful - running the actual
installation is the easy part; determining what to download from PyPI
(or other sources) looks much harder.