Add bootstrap preparation logic
A separate script is used to create the bootstrap/ directory from
scratch. Aside from the vendored dependency wheels, the core wheels are
also built and put directly into the bootstrap, using the `build` API.

The `make-zipapp` script is simplified; since the core wheels will
already be in the bootstrap folder, they don't need to be located in the
per-project 'dist' directories.

The `bootstrap/` directory is no longer included in the repository,
since the vendored wheels can now be obtained on demand.

Currently, wheels are obtained using the `download` functionality in
`rfbs.fetch`. It should instead try to obtain them from cache first, so
that a locally installed PAPER can power the creation of a new zipapp
without an Internet connection (as long as the needed wheels are cached).