Improve wheel tag handling
Since the wheels to install in the bootstrap use both `py2.py3` and
`py3` Python wheel tags, and the bootstrap installation now uses this
API, it's important to ensure both are supported.

For now, attempts to find a wheel in the cache just try both versions of
the wheel tags and report the first success; attempts to download to
cache try both wheel tags independently (so a cache could potentially
end up with multiple wheels for the same name and version).

Note that the downloading code here isn't really tested. It also simply
ignores any download errors.

Going forward, logic from `packaging` should be used to determine what
wheel tag-sets to use and in what order.