Implement basic setup from unpacked cache
Unpacked wheels in the cache can now be set up in an environment, by
hard-linking the files (and making corresponding folders).

This is sufficient for many wheels; but it skips creating console
scripts (which will soon be needed for paper-ui). Further, the
.dist-info should probably be copied rather than being hard-linked, and
somewhere along the line RECORD should be updated. There's also no
provision for pre-compiling to .pyc yet, or spreading contents from
.data folders.

This serves as a demonstration of the _clone code, and one of the
reasons not to use `installer` - it didn't appear to be designed to
support manipulating the copy process at this level.