Bugfix moving files from temp dir
Apparently, `shutil.move` of one folder into another will not create a
subfolder with the source folder's name, but instead "promote" its
contents directly into the destination.

This is avoided by explicitly adding the source's name to the `move`
call in the loop. This should also work for files - although neither
wheels nor sdists are supposed to have top-level files.