Fix installation procedure
The zipapp's bootstrap installer is fixed to match other recent changes.
It unpacks the bootstrap wheels into a temporary directory, puts all of
them on `sys.path`, then uses its own code to do a proper installation -
first copying the wheels into cache, then installing from the cache into
a virtual environment.

A diagnostic message is also added at the start of the process to
indicate the base version of Python used for the installation. This will
also be the Python used to run PAPER commands later.